Fix naming of NetCDF files, and baseline for output NetCDF file, so that plot set 3 will test ok.
