+++ test/baselines/plotset3/figure-set3_Global_JJA_T_plot-combined.nc
